+++ src/advanced_search_module.cpp
@@ -6,6 +6,6 @@ namespace py = pybind11;
 
 PYBIND11_MODULE(advanced_search_cpp, m) {
     py::class_<AdvancedSearch>(m, "AdvancedSearch")
-        .def(py::init<py::array_t<double>>())
+        .def(py::init<py::array_t<float>>())
         .def("search", &AdvancedSearch::search);
 }
